Minor Audit changes

+++ contracts/TestERC20.sol
@@ -5,7 +5,7 @@ import "openzeppelin-solidity-2.3.0/contracts/token/ERC20/ERC20Detailed.sol";
 import "openzeppelin-solidity-2.3.0/contracts/token/ERC20/ERC20Mintable.sol";
 
 contract TestERC20 is ERC20Detailed, ERC20Mintable {
-    constructor(uint amount) ERC20Detailed('Test ERC20', 'TEST', 18) public {
+    constructor(uint256 amount) ERC20Detailed('Test ERC20', 'TEST', 18) public {
         mint(msg.sender, amount);
     }
 }
